Re-enable test store.wat
Test was to slow because once the out of bounds cases have been ruled
out, all "in bounds" cases still have to concretize a value for the symbol
and perform the store, leading to approx. <page_size> calls to Z3.

By reducing the possible values of idx, we ensure that the test runs fast.