In tests, change dune exec for direct binary call
Using `dune exec`` means that the dune-project file
needs to be present in the dune sandbox, which
would sometimes not happen and lead to the very
annoying workflow of having to dune clean before
every dune runtest.

I tried adding the dune-project file as a dependency
but couldn't get it to work.

A direct call to the binary seems to solve the issue.