Add breadcrumbs to Thread.t
As identified in https://github.com/OCamlPro/owi/issues/208, the
deterministic order option is not deterministic across architectures.

A likely culprit is the use of the polymorphic compare function which
may not give the same result in different environments or on different
targets.

This commit resolves this by keeping track of each choice made ('select')
through "breadcrumbs" and sorting results on these breadcrumbs.

Closes: https://github.com/OCamlPro/owi/issues/208